requirements_airflow.txt
# Airflow and Database Requirements - Latest Compatible Versions
apache-airflow
apache-airflow-providers-postgres
psycopg2-binary
sqlalchemy

# For data processing
pandas
numpy
